refactor: improve structure
