+++ lib/Compilation.js
@@ -28,11 +28,6 @@ const DependencyTemplates = require("./DependencyTemplates");
 const Entrypoint = require("./Entrypoint");
 const ErrorHelpers = require("./ErrorHelpers");
 const FileSystemInfo = require("./FileSystemInfo");
-const {
-	connectChunkGroupAndChunk,
-	connectChunkGroupParentAndChild,
-	connectEntrypointAndDependOn
-} = require("./GraphHelpers");
 const MainTemplate = require("./MainTemplate");
 const Module = require("./Module");
 const ModuleGraph = require("./ModuleGraph");
@@ -3312,7 +3307,10 @@ BREAKING CHANGE: Asset processing hooks in Compilation has been merged into a si
 			this.namedChunkGroups.set(name, entrypoint);
 			this.entrypoints.set(name, entrypoint);
 			this.chunkGroups.push(entrypoint);
-			connectChunkGroupAndChunk(entrypoint, chunk);
+
+			if (entrypoint.pushChunk(chunk)) {
+				chunk.addGroup(entrypoint);
+			}
 
 			/** @type {Set<Module>} */
 			const entryModules = new Set();
@@ -3403,8 +3401,12 @@ Remove the 'runtime' option from the entrypoint.`);
 						entry.setRuntimeChunk(entryChunk);
 						continue outer;
 					}
-					connectEntrypointAndDependOn(entry, dependency);
-					connectChunkGroupParentAndChild(dependency, entry);
+
+					entry.addDependOn(dependency);
+
+					if (dependency.addChild(entry)) {
+						entry.addParent(dependency);
+					}
 				}
 			} else if (runtime) {
 				const entry = /** @type {Entrypoint} */ (this.entrypoints.get(name));
@@ -4156,7 +4158,9 @@ Or do you want to use the entrypoints '${name}' and '${runtime}' independently o
 		}
 		const chunk = this.addChunk(name);
 
-		connectChunkGroupAndChunk(chunkGroup, chunk);
+		if (chunkGroup.pushChunk(chunk)) {
+			chunk.addGroup(chunkGroup);
+		}
 
 		this.chunkGroups.push(chunkGroup);
 		if (name) {
@@ -4200,7 +4204,9 @@ Or do you want to use the entrypoints '${name}' and '${runtime}' independently o
 		}
 		this.chunkGroups.push(entrypoint);
 		this.asyncEntrypoints.push(entrypoint);
-		connectChunkGroupAndChunk(entrypoint, chunk);
+		if (entrypoint.pushChunk(chunk)) {
+			chunk.addGroup(entrypoint);
+		}
 		if (module) {
 			entrypoint.addOrigin(module, loc, request);
 		}
@@ -5493,7 +5499,9 @@ This prevents using hashes of each other and should be avoided.`);
 					...options.entryOptions
 				});
 				chunkGraph.connectChunkAndEntryModule(chunk, module, entrypoint);
-				connectChunkGroupAndChunk(entrypoint, chunk);
+				if (entrypoint.pushChunk(chunk)) {
+					chunk.addGroup(entrypoint);
+				}
 				entrypoint.setRuntimeChunk(chunk);
 				entrypoint.setEntrypointChunk(chunk);
