+++ lib/NormalModule.js
@@ -31,7 +31,7 @@ const ModuleParseError = require("./errors/ModuleParseError");
 const ModuleWarning = require("./errors/ModuleWarning");
 const UnhandledSchemeError = require("./errors/UnhandledSchemeError");
 const WebpackError = require("./errors/WebpackError");
-const formatLocation = require("./formatLocation");
+const formatLocation = require("./util/formatLocation");
 const LazySet = require("./util/LazySet");
 const { isSubset } = require("./util/SetHelpers");
 const { getScheme } = require("./util/URLAbsoluteSpecifier");
