+++ lib/WebpackOptionsApply.js
@@ -517,7 +517,7 @@ class WebpackOptionsApply extends OptionsApply {
 		}).apply(compiler);
 		if (options.amd !== false) {
 			const AMDPlugin = require("./dependencies/AMDPlugin");
-			const RequireJsStuffPlugin = require("./RequireJsStuffPlugin");
+			const RequireJsStuffPlugin = require("./dependencies/RequireJsStuffPlugin");
 
 			new AMDPlugin(options.amd || {}).apply(compiler);
 			new RequireJsStuffPlugin().apply(compiler);
