+++ lib/buildChunkGraph.js
@@ -5,7 +5,6 @@
 
 "use strict";
 
-const { connectChunkGroupParentAndChild } = require("./GraphHelpers");
 const ModuleGraphConnection = require("./ModuleGraphConnection");
 const AsyncDependencyToInitialChunkError = require("./errors/AsyncDependencyToInitialChunkError");
 const { getEntryRuntime, mergeRuntime } = require("./util/runtime");
@@ -1311,10 +1310,9 @@ const connectChunkGroups = (
 			chunkGraph.connectBlockAndChunkGroup(block, chunkGroup);
 
 			// 4. Connect chunk with parent
-			connectChunkGroupParentAndChild(
-				originChunkGroupInfo.chunkGroup,
-				chunkGroup
-			);
+			if (originChunkGroupInfo.chunkGroup.addChild(chunkGroup)) {
+				chunkGroup.addParent(originChunkGroupInfo.chunkGroup);
+			}
 		}
 	}
 };
