+++ lib/cache/PackFileCacheStrategy.js
@@ -7,7 +7,7 @@
 
 const FileSystemInfo = require("../FileSystemInfo");
 const ProgressPlugin = require("../ProgressPlugin");
-const { formatSize } = require("../SizeFormatHelpers");
+const formatSize = require("../util/formatSize");
 const SerializerMiddleware = require("../serialization/SerializerMiddleware");
 const LazySet = require("../util/LazySet");
 const makeSerializable = require("../util/makeSerializable");
