+++ lib/css/CssModule.js
@@ -5,15 +5,15 @@
 
 "use strict";
 
-const NormalModule = require("./NormalModule");
-const makeSerializable = require("./util/makeSerializable");
+const NormalModule = require("../NormalModule");
+const makeSerializable = require("../util/makeSerializable");
 
-/** @typedef {import("./Module")} Module */
-/** @typedef {import("./NormalModule").NormalModuleCreateData} NormalModuleCreateData */
-/** @typedef {import("./RequestShortener")} RequestShortener */
-/** @typedef {import("./serialization/ObjectMiddleware").ObjectDeserializerContext} ObjectDeserializerContext */
-/** @typedef {import("./serialization/ObjectMiddleware").ObjectSerializerContext} ObjectSerializerContext */
-/** @typedef {import("../declarations/WebpackOptions").CssParserExportType} CssParserExportType */
+/** @typedef {import("../Module")} Module */
+/** @typedef {import("../NormalModule").NormalModuleCreateData} NormalModuleCreateData */
+/** @typedef {import("../RequestShortener")} RequestShortener */
+/** @typedef {import("../serialization/ObjectMiddleware").ObjectDeserializerContext} ObjectDeserializerContext */
+/** @typedef {import("../serialization/ObjectMiddleware").ObjectSerializerContext} ObjectSerializerContext */
+/** @typedef {import("../../declarations/WebpackOptions").CssParserExportType} CssParserExportType */
 
 /** @typedef {string | undefined} CssLayer */
 /** @typedef {string | undefined} Supports */
