+++ lib/dependencies/CommonJsExportsParserPlugin.js
@@ -6,7 +6,7 @@
 "use strict";
 
 const RuntimeGlobals = require("../RuntimeGlobals");
-const formatLocation = require("../formatLocation");
+const formatLocation = require("../util/formatLocation");
 const { evaluateToString } = require("../javascript/JavascriptParserHelpers");
 const { propertyAccess } = require("../util/property");
 const CommonJsExportRequireDependency = require("./CommonJsExportRequireDependency");
