+++ lib/dependencies/CssIcssImportDependency.js
@@ -13,7 +13,7 @@ const CssImportDependency = require("./CssImportDependency");
 /** @typedef {import("../Dependency")} Dependency */
 /** @typedef {import("../DependencyTemplate").CssDependencyTemplateContext} DependencyTemplateContext */
 /** @typedef {import("../Module")} Module */
-/** @typedef {import("../CssModule")} CssModule */
+/** @typedef {import("../css/CssModule")} CssModule */
 /** @typedef {import("../ModuleGraph")} ModuleGraph */
 /** @typedef {import("../util/runtime").RuntimeSpec} RuntimeSpec */
 /** @typedef {import("../javascript/JavascriptParser").Range} Range */
