+++ lib/dependencies/RequireJsStuffPlugin.js
@@ -8,16 +8,16 @@
 const {
 	JAVASCRIPT_MODULE_TYPE_AUTO,
 	JAVASCRIPT_MODULE_TYPE_DYNAMIC
-} = require("./ModuleTypeConstants");
-const RuntimeGlobals = require("./RuntimeGlobals");
-const ConstDependency = require("./dependencies/ConstDependency");
+} = require("../ModuleTypeConstants");
+const RuntimeGlobals = require("../RuntimeGlobals");
 const {
 	toConstantDependency
-} = require("./javascript/JavascriptParserHelpers");
+} = require("../javascript/JavascriptParserHelpers");
+const ConstDependency = require("./ConstDependency");
 
-/** @typedef {import("../declarations/WebpackOptions").JavascriptParserOptions} JavascriptParserOptions */
-/** @typedef {import("./Compiler")} Compiler */
-/** @typedef {import("./javascript/JavascriptParser")} JavascriptParser */
+/** @typedef {import("../../declarations/WebpackOptions").JavascriptParserOptions} JavascriptParserOptions */
+/** @typedef {import("../Compiler")} Compiler */
+/** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
 
 const PLUGIN_NAME = "RequireJsStuffPlugin";
