+++ lib/optimize/MinMaxSizeWarning.js
@@ -5,8 +5,8 @@
 
 "use strict";
 
-const SizeFormatHelpers = require("../SizeFormatHelpers");
 const WebpackError = require("../errors/WebpackError");
+const formatSize = require("../util/formatSize");
 
 class MinMaxSizeWarning extends WebpackError {
 	/**
@@ -26,8 +26,8 @@ class MinMaxSizeWarning extends WebpackError {
 		super(
 			"SplitChunksPlugin\n" +
 				`${keysMessage}\n` +
-				`Configured minSize (${SizeFormatHelpers.formatSize(minSize)}) is ` +
-				`bigger than maxSize (${SizeFormatHelpers.formatSize(maxSize)}).\n` +
+				`Configured minSize (${formatSize(minSize)}) is ` +
+				`bigger than maxSize (${formatSize(maxSize)}).\n` +
 				"This seem to be a invalid optimization.splitChunks configuration."
 		);
 	}
