+++ lib/optimize/SideEffectsFlagPlugin.js
@@ -14,7 +14,7 @@ const {
 const { STAGE_DEFAULT } = require("../OptimizationStages");
 const HarmonyExportImportedSpecifierDependency = require("../dependencies/HarmonyExportImportedSpecifierDependency");
 const HarmonyImportSpecifierDependency = require("../dependencies/HarmonyImportSpecifierDependency");
-const formatLocation = require("../formatLocation");
+const formatLocation = require("../util/formatLocation");
 const { CompilerHintNotationRegExp } = require("../util/magicComment");
 
 /** @typedef {import("estree").MaybeNamedClassDeclaration} MaybeNamedClassDeclaration */
