+++ lib/performance/EntrypointsOverSizeLimitWarning.js
@@ -5,8 +5,8 @@
 
 "use strict";
 
-const { formatSize } = require("../SizeFormatHelpers");
 const WebpackError = require("../errors/WebpackError");
+const formatSize = require("../util/formatSize");
 
 /** @typedef {import("./SizeLimitsPlugin").EntrypointDetails} EntrypointDetails */
