+++ lib/stats/DefaultStatsFactoryPlugin.js
@@ -8,7 +8,7 @@
 const util = require("util");
 const { WEBPACK_MODULE_TYPE_RUNTIME } = require("../ModuleTypeConstants");
 const ModuleDependency = require("../dependencies/ModuleDependency");
-const formatLocation = require("../formatLocation");
+const formatLocation = require("../util/formatLocation");
 const { LogType } = require("../logging/Logger");
 const AggressiveSplittingPlugin = require("../optimize/AggressiveSplittingPlugin");
 const SizeLimitsPlugin = require("../performance/SizeLimitsPlugin");
