+++ lib/stats/DefaultStatsPrinterPlugin.js
@@ -1510,7 +1510,7 @@ const AVAILABLE_FORMATS = {
 		(oversize ? yellow : green)(filename),
 	formatFlag: (flag) => `[${flag}]`,
 	formatLayer: (layer) => `(in ${layer})`,
-	formatSize: require("../SizeFormatHelpers").formatSize,
+	formatSize: require("../util/formatSize"),
 	formatDateTime: (dateTime, { bold }) => {
 		const d = new Date(dateTime);
 		const x = twoDigit;
