+++ lib/util/formatLocation.js
@@ -5,8 +5,8 @@
 
 "use strict";
 
-/** @typedef {import("./Dependency").DependencyLocation} DependencyLocation */
-/** @typedef {import("./Dependency").SourcePosition} SourcePosition */
+/** @typedef {import("../Dependency").DependencyLocation} DependencyLocation */
+/** @typedef {import("../Dependency").SourcePosition} SourcePosition */
 
 /**
  * Returns formatted position.
