+++ lib/util/formatSize.js
@@ -10,7 +10,7 @@
  * @param {number=} size the size in bytes
  * @returns {string} the formatted size
  */
-module.exports.formatSize = (size) => {
+const formatSize = (size) => {
 	if (typeof size !== "number" || Number.isNaN(size) === true) {
 		return "unknown size";
 	}
@@ -24,3 +24,5 @@ module.exports.formatSize = (size) => {
 
 	return `${Number((size / 1024 ** index).toPrecision(3))} ${abbreviations[index]}`;
 };
+
+module.exports = formatSize;
