+++ lib/util/internalSerializables.js
@@ -185,7 +185,7 @@ module.exports = {
 	InitFragment: () => require("../InitFragment"),
 	ModuleGraph: () => require("../ModuleGraph"),
 	NormalModule: () => require("../NormalModule"),
-	CssModule: () => require("../CssModule"),
+	CssModule: () => require("../css/CssModule"),
 	RawDataUrlModule: () => require("../asset/RawDataUrlModule"),
 	RawModule: () => require("../RawModule"),
 	"sharing/ConsumeSharedModule": () =>
