+++ lib/wasm-sync/WasmFinalizeExportsPlugin.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-const formatLocation = require("../formatLocation");
+const formatLocation = require("../util/formatLocation");
 const UnsupportedWebAssemblyFeatureError = require("./UnsupportedWebAssemblyFeatureError");
 
 /** @typedef {import("../Compiler")} Compiler */
